--- lib/colors.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-const kColorDark = Color(0xff333333);
-const kColorAppBarFont = Colors.white;
-const kBackgroundColor = Color(0xffCDF0EA);
-const kColorTextField = Color(0xffFBF7F0);
-const kTextBubbleColor = Color(0xff395B64);
