+++ lib/main.dart
@@ -1,6 +1,5 @@
-import 'package:chat_app/chat_screen.dart';
-import 'package:chat_app/home_screen.dart';
-import 'package:chat_app/login_screen.dart';
+import 'package:chat_app/screens/home_screen.dart';
+import 'package:chat_app/screens/login_screen.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
