+++ lib/resources/colors.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+const kColorDark = Color(0xff333333);
+const kColorLight = Colors.white;
+const kThemeColor = Color(0xffCDF0EA);
+const kColorTextField = Color(0xffFBF7F0);
+const kTextBubbleColor = Color(0xff395B64);
+const kColorHintText = Color(0xff596E79);
