+++ lib/resources/styles.dart
@@ -1,0 +1,6 @@
+import 'package:chat_app/resources/colors.dart';
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+var kFontStyleHeadingLarge = GoogleFonts.lexendDeca(textStyle: TextStyle(color: kThemeColor, fontSize: 60));
+var kFontStyleHeadingMedium = GoogleFonts.lexendDeca(textStyle: TextStyle(color: kThemeColor, fontSize: 40));
